+++ image_generator/__init__.py
@@ -27,10 +27,11 @@ https://github.com/TheElementalOfDestruction/image-generator
 #    along with this program. If not, see <http://www.gnu.org/licenses/>.
 
 __author__ = 'Destiny Peterson (The Elemental of Destruction)'
-__date__ = '2021-05-26'
-__version__ = '1.2.0'
+__date__ = '2021-05-30'
+__version__ = '1.3.0'
 
 from image_generator.drake import createDrake
+from image_generator.realization import createRealization
 from image_generator.ship import createShip
 from image_generator.test_font import createFontTest
 from image_generator.trash import createTrash