+++ image_generator/drake/__init__.py
@@ -1,1 +1,1 @@
-from image_generator.drake.builder import createDrake
+from image_generator.drake.core import createDrake